feat: 시연 직전 폴리시 묶음 — 사진 로드 fix + 프로젝트명 순번 + UI 다듬기
- ProjectPage 사진 박스: setImageURI를 Glide로 교체, ngrok-skip-browser-warning과
  Authorization 헤더를 GlideUrl LazyHeaders로 첨부 (HTTPS 사진이 흰/노랑 박스로 뜨던 문제)
- 백엔드 image_url이 상대 경로일 가능성 대비 toAbsoluteImageUrl로 BASE_URL 결합
- Glide RequestListener 로 onLoadFailed/onResourceReady 진단 로그 추가
- 세부사항 6개 필드 폭을 wrap_content + minWidth로 변경해 텍스트 길이에 맞춰 자동 조정
- 차시별 그래프 X축에 양쪽 padding(-0.35 ~ N-0.65) 추가해 데이터 점이 Y축에 안 닿게
- 수수료 소수점 절삭 제거: Platform.defaultFee int→double, SimulationData.platformFee Integer→Double, formatFee 헬퍼
- 프로젝트명을 GET /projects 카운트 기반 "프로젝트N" 자동 부여 (긴 이름 대신 깔끔하게)
- 기타: RetrofitClient에 ngrok 헤더, StatisticsFragment status default→planning,
  SimulationData.reset()/clearPostNavigation() 분리, EditText 콤마 포맷·단위 추가

+++ app/src/main/java/com/example/creator_flow/model/SimulationData.java
@@ -41,7 +41,7 @@ public class SimulationData {
 
     // ===== Platform =====
     public static String platformName;     // 판매 업체 (윗치폼/텀블벅 등) — 표시용
-    public static Integer platformFee;     // 수수료(%) — Business일 때만 의미 있음
+    public static Double platformFee;      // 수수료(%) — 소수점 유지 (예: 4.68)
     /** POST /simulations 에 사용할 platform_plan의 실제 UUID. mock 모드에선 null */
     public static String platformPlanId;
